Lyric_backend/requirements.txt
fastapi
langchain
langchain_openai
python-dotenv
uvicorn
fastapi-cors
